4.21 tehty

+++ osa4/controllers/blogs.js
@@ -37,6 +37,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+
+	// eslint-disable-next-line no-undef
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	if (!request.token || !decodedToken.id) {
+		return response.status(401).json({ error: 'token missing or invalid' })
+	}
+
+	const user = await User.findById(decodedToken.id)
+	const blog = await Blog.findById(request.params.id)
+	const blogUser = blog.user._id.toString()
+
+	if(user.id !== blogUser) {
+		return response.status(401).json({ error: 'wrong user' })
+	}
+
 	await Blog.findByIdAndRemove(request.params.id)
 	response.status(204).end()
 })
